imp: tracks ibc-rs pr915
